+++ src/app/layout.tsx
@@ -1,6 +1,7 @@
+import "@radix-ui/themes/styles.css";
+import { Theme } from "@radix-ui/themes";
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
 
 const geistSans = Geist({
 	variable: "--font-geist-sans",
@@ -25,7 +26,7 @@ export default function RootLayout({
 	return (
 		<html lang="en">
 			<body className={`${geistSans.variable} ${geistMono.variable}`}>
-				{children}
+				<Theme>{children}</Theme>
 			</body>
 		</html>
 	);
